feat:Implement manifest schemata and add version checking to client
Added versioning information to the main and build manifests. Added some pre-build scripts to generate compile-time constants for version numbers, for use across the CLI and client. the client will now reject wrappers which do not target the runtime version provided by the client (this is the default, but the restriction can be lowered to simply warn to `console.error` by changing the client config)